Fix .data region origin for newer devices
For new devices, Microchip decided to change the way the data region
origin is defined in the device-specs. This patch fixes the linker
script template, to use the value from a symbol exported in the crt*
libraries (included in the atpacks), if defined.

This patch is based on another patch from the arduino/toolchian-avr
repository:

https://github.com/arduino/toolchain-avr/blob/26d5bd033241d834c9c01075aa8ea17dda62c4ca/binutils-patches/00-binutils-data_region_length.patch